added example files
Three example file showing basic usage are added to examples/:
local.config
local-to-ssh.config
ssh-to-local.config

As filenames suggest, usages are local-only (1) and local-remote
scenarios. Note that NFS backups are regarded to belong to category (1),
the local backup.

Example files show minimal requirements for parameters. Note that the
USE_SSH options is somewhat redundant and subject to removal in later
versions.